Create Day 1:Data Types
Task
Complete the code in the editor below. The variables i,d , and s are already declared and initialized for you. You must:
1.Declare 3 variables: one of type int, one of type double, and one of type String.
2.Read 3 lines of input from stdin (according to the sequence given in the Input Format section below) and initialize your 3 variables.
3.Use the + operator to perform the following operations:
a)Print the sum of i plus your int variable on a new line.
b)Print the sum of d plus your double variable to a scale of one decimal place on a new line.
c)Concatenate s with the string you read as input and print the result on a new line.